chore: cleanup fine-grained example script
